Updated sizing and some refactoring
