+++ frontend/libs/studio-feedback-form/src/FeedbackForm/Question/TextQuestion.tsx
@@ -8,6 +8,7 @@ export function TextQuestion({ id, label, value, onChange }: QuestionsProps): Re
   const debouncedOnChange = (newValue: string) => debounce(() => onChange(id, newValue));
   return (
     <StudioTextarea
+      size='sm'
       id={id}
       label={label}
       value={value}
