+++ frontend/packages/ux-editor/src/containers/DesignView/AddItem/ToggleAddComponentPoc.module.css
@@ -1,0 +1,5 @@
+.switchWrapper {
+  display: flex;
+  flex-direction: row;
+  gap: var(--fds-spacing-2);
+}
